Check if a variable is set with the php function isset($var).
Because the stdClass is dynamic, a variable can be avialbe, but must not be. Insteand of just checking if aviable (by accessing it) it's much more better to aks if the variable is realy set. Why? In development mode on php, you will get thousends of warnings (ok, there are 4) that a var is not set in this content.
So the parser can easier check itself (bool-check) than type-check to null.