Fixes broken changePassword method for SagUserUtils
This commit fixes a couple issues.

1) SagUserUtils->changePassword() was not working. It was recording a
new SHA1 without also updating the salt used for the new password. It
would then update the password but the user could never log in again as
the salt didn’t match the new SHA1 hash.

2) The documentation stated that SagUserUtils->getUser() is what $doc
was expected to be like. If $doc was passed in from that method
however, it did not work. It needs the ->body property from the
returned object.